@@ -1,9 +1,9 @@
 import { Injectable } from '@angular/core';
-import { CanActivate, Router, CanActivateChild} from '@angular/router';
+import { CanActivate, Router, CanActivateChild, Route, CanLoad} from '@angular/router';
 import { AuthService } from './auth.service';
 
 @Injectable()
-export class AuthGuard implements CanActivate, CanActivateChild {
+export class AuthGuard implements CanActivate, CanActivateChild, CanLoad {
 
   constructor(private authService: AuthService, private router: Router) { }
 
@@ -13,6 +13,11 @@ export class AuthGuard implements CanActivate, CanActivateChild {
   }
 
   canActivateChild(){
+    if(this.authService.isLoggedIn()) return true;
+    this.router.navigate(['/home']);
+  }
+
+  canLoad(): boolean {
     if(this.authService.isLoggedIn()) return true;
     this.router.navigate(['/home']);
   }
